Rename generated_data.[ch] to generated_code.[ch]
After renaming generated_code.c, update #include.

In launch.json and tasks.json, change program name from
"generated_data" to "daffodil".

In .gitignore, change pattern from generated_data to daffodil.
