+++ daffodil-runtime2/src/main/c/.vscode/launch.json
@@ -9,7 +9,7 @@
             "name": "gcc - Build and debug active file",
             "type": "cppdbg",
             "request": "launch",
-            "program": "${fileDirname}/generated_data",
+            "program": "${fileDirname}/daffodil",
             "args": ["unparse", "test_unparse_input"],
             "stopAtEntry": false,
             "cwd": "${workspaceFolder}",
